- move source components to source folder - separate functions in collaborative editor in a new file - create new sourceListItem component
